add animationRenderer

+++ app/src/main/java/com/example/glesgameengine/GraphicSystem/RenderTarget.java
@@ -5,6 +5,6 @@ import java.nio.FloatBuffer;
 public class RenderTarget
 {
     public FloatBuffer matrix;
-    public float z_index;
+    public int z_index;
     public int imageCode;
 }
